New dummy tidal model
As discussed, we adapt this approach for activation of coordinates, instead of modifying the core of the code. In this way, when tides are switched on "True", then whatever the objects that are not "tidally on" (tidally perturbed), the object should choose DisabledModel instead of the Disabled features in under "tides".

I included also three examples for generating .json for related test cases, they should serve as examples for the usage of the new tidal model.