Update of the Python part of the test cases
I include also the Python part of the test cases (`pytest`, mainly for Kaula), notices that the Kaula tests do not pass the `pytest` originally, due to the comparison between values of "-0.0" and "0.0" under the same field.

I added a function called "normalize_json" to transform the "-0.0" to "0.0" under the same field within .json for the comparison, as we know they are practically the same.

Also I improve a tiny bit of the integrity of the `load_kaula_parameters` function for Rust tests.