Physics spectrum rescaling (#1)
Due to the feature of stellar tides, we need to take into account the real-time stellar spin rate into the love number extracted for simulation, thus the modification in kaula.rs. Now the code can tell whether we are working on stellar/planetary tide on the Kaula model.

Allow bodies to switch on/off the tides by using DisabledModel, but only in combination with Kaula model.
Notice that the combination of CTL model and DisabledModel does not work at the same time, due to the special structure in CTL model.


* Minor bug-fix for true_longitude

Minor bug fix for calculation of `true_longitude` in the function `_calculate_keplerian_orbital_elements`.

* Minor update: Bug on reading input file

In running normal simulations, the code would show the following error:

`FileNotFoundError: /opt/homebrew/lib/python3.11/site-packages/posidoniusinput/Leconte_2011/mass_80.0000.dat not found.`

The correct path should be `posidonius/input`, I fixed the `BASE_DIR` in this way, such that the code can read the input file, for example for stellar evolution.

* Add particle.spin for checking in test cases

I modified `test/common/universe.rs`, so that it is going to check the evolution of spin as well. One can see the value of position/velocity/acceleration did not change at all, so the values remain valid for future code-checking.

Notice I excluded the Kaula part for the moment, as the test requires some other debugging first.

RMK: Several `rust_vs_python` test cases do not pass, including:
test_evolution-solar_like_baraffe2015_rust_vs_python
test_evolution-solar_like_bolmontmathis2016_rust_vs_python
test_evolution-solar_like_non_evolving_rust_vs_python
ias15_rust_vs_python

See my next commit for detail.

* Modification of spin test

Several tests failed before (see previous commit) due to the different in .json (python and rust), we lower the precision for the spin test as we know now, the differences in output is acceptable since the .json are different.

* Update on the computation of the love number arrasys

Unifying the ways of love number computation between rust and python, by extending the array using the final elements instead of adding zeros at front.

Rmk: This discrepancy cause certain tests to fail, and we might want to consider sorting out the data first before doing binary-tree search in `kaula.rs`.

* New dummy tidal model

As discussed, we adapt this approach for activation of coordinates, instead of modifying the core of the code. In this way, when tides are switched on "True", then whatever the objects that are not "tidally on" (tidally perturbed), the object should choose DisabledModel instead of the Disabled features in under "tides".

I included also three examples for generating .json for related test cases, they should serve as examples for the usage of the new tidal model.

* Test cases update specific for kaula model

.json under posidonius/tests/data/test_tides_kaula-xyz: for the xyz_rust_vs_python test cases, generated by the python scripts from the previous commit.

{test_tides_kaula, stars, planets}.rs under tests/: the new test cases for testing kaula. Notice these tests are definitely not optimized as in I created several extra function just for my specific kaula test.
On top of that, I do not include other physical mechanisms in the same tests, so there could be some conflicts between models. These tests should be updated and modified in a neat way in the future.

I added also the .json of expected result for these new 3 sets of test cases: as they passed the rust_vs_python tests, meaning the simulation results from .json generated between rust and python agree on each other, thus the results from these json should be trustable.

* Upload of scientific test cases

Upload of 5 scientific test cases for kaula model, so that one could check the expected behaviour from Kaula model, such as the spin-orbit resonance for planets.

* Update of the Python part of the test cases

I include also the Python part of the test cases (`pytest`, mainly for Kaula), notices that the Kaula tests do not pass the `pytest` originally, due to the comparison between values of "-0.0" and "0.0" under the same field.

I added a function called "normalize_json" to transform the "-0.0" to "0.0" under the same field within .json for the comparison, as we know they are practically the same.

Also I improve a tiny bit of the integrity of the `load_kaula_parameters` function for Rust tests.

* Update of several Python scripts

Modification of several Python scripts generating .json, for the integrity of reading the love number data.

* Typo in python scripts

Correct certain typo in the python scripts

* Rescaling of the love number

Modification taking into account of the real time evolution of the stellar spin rate, rescaling the love number spectrum on both x and y axis.

Practically speaking, we only modified the tidal frequency, and rescale the love number k2 read, instead of modifying the spectrum every time step.

Also, for the moment I don't know how to translate "true" between python and rust, thus I use integer -1 and 1 to do the job.

* Upload for several scientific test cases

Several scientific test cases of single planet migration.

* Update on the previous test cases

Update the python scripts otherwise it won't work under the new implementation.

* Resonant states construction

Python script to generate 2:1 MMR from two circularized orbit.

* Update script naming

Correct (switching) the naming

* Upload of Python scripts

* Upload of a old python scripts for checking

* Update the resonance analysis script

Issues for explore_timed_resonances mostly, as some parameters need to be integer

* Grouping of scripts of simulations

These are the scripts for the comparison between the kaula model and CTL model.

* Scripts for creating the 2:1 MMR

* Scripts for 2:1 MMR

Scripts for testing the stability of those created 2:1 MMR.

* Upload of simulation scripts

These scripts are for studying the influence of the stellar tide on the MMR.

* Scripts correction

We decided to switch to WHFast for the simulations, thus for the update.

* Scripts upload

Simulation scripts for a05.

* Correction for the previous uploaded scripts.

* Upload of simulation scripts for a07

* Remove useless simulation scripts

* Minor update for the time limit for simulation scripts

* Update the necessary module for Rust test

* Update related to the python and rust test cases

For the stars.py and planets.py, I added the necessary structure for generating the correct python test cases.

From the mod.rs, the new structural changes are there already. For the new Rust .json, one needs to remove the old one and generate the new ones simply but running the Rust tests.

* Test cases update adapting the new Kaula parameter structure

We use the stellar spectrum "alpha0.51600_P1p2_Ek1p5em6.txt" for all test cases of stellar tides (kaula).